requirements.txt
# OpenRouterBench Core Dependencies
# For data_collector, evaluation, and generators modules
# Note: Each baseline has its own requirements.txt due to version conflicts

# LLM & API
openai>=1.0.0
tiktoken>=0.4.0
httpx>=0.24.0
litellm>=1.50.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
datasets>=2.0.0

# ML & Evaluation
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.0
sympy>=1.12
nltk>=3.8.0
pylatexenc>=2.10

# Utilities
loguru>=0.6.0
pyyaml>=6.0
tqdm>=4.62.0
tenacity>=8.0.0
python-dotenv>=1.0.0

# Caching
pymysql>=1.0.0
